allez là

+++ meuporg-server/source/Components/BasicStatsComponent.cpp
@@ -1,6 +1,6 @@
 #include "BasicStatsComponent.hpp"
 
-BasicStatsComponent::BasicStatsComponent(unsigned long long ownerId)
+BasicStatsComponent::BasicStatsComponent(unsigned int ownerId)
      : kantan::Component(std::string("BasicStats"), ownerId)
      , isDead(false)
      , hp(1.f)
