+++ meuporg-server/source/Components/ClientLinkComponent.cpp
@@ -1,6 +1,6 @@
 #include "ClientLinkComponent.hpp"
 
-ClientLinkComponent::ClientLinkComponent(unsigned long long ownerId)
+ClientLinkComponent::ClientLinkComponent(unsigned int ownerId)
      : kantan::Component(std::string("ClientLink"), ownerId)
      , client(nullptr)
      , dbid(-1)
