+++ meuporg-server/source/Components/DamageComponent.cpp
@@ -1,6 +1,6 @@
 #include "DamageComponent.hpp"
 
-DamageComponent::DamageComponent(unsigned long long ownerId)
+DamageComponent::DamageComponent(unsigned int ownerId)
      : kantan::Component(std::string("Damage"), ownerId)
 	 , emitter(-1)
 	 , damage(1.f)
