+++ meuporg-server/source/Components/LevelStatsComponent.cpp
@@ -1,6 +1,6 @@
 #include "LevelStatsComponent.hpp"
 
-LevelStatsComponent::LevelStatsComponent(unsigned long long ownerId)
+LevelStatsComponent::LevelStatsComponent(unsigned int ownerId)
      : kantan::Component(std::string("LevelStats"), ownerId)
      , xp(0.f)
 	 , level(0.f)
