+++ meuporg-server/source/Components/MonsterAIComponent.cpp
@@ -1,6 +1,6 @@
 #include "MonsterAIComponent.hpp"
 
-MonsterAIComponent::MonsterAIComponent(unsigned long long ownerId)
+MonsterAIComponent::MonsterAIComponent(unsigned int ownerId)
      : kantan::Component(std::string("MonsterAI"), ownerId)
      , direction(1)
      , lastDecision(sf::Time::Zero)
