+++ meuporg-server/source/Components/NameComponent.cpp
@@ -1,6 +1,6 @@
 #include "NameComponent.hpp"
 
-NameComponent::NameComponent(unsigned long long ownerId)
+NameComponent::NameComponent(unsigned int ownerId)
      : kantan::Component(std::string("Name"), ownerId)
      , id(-1)
      , name("#")
