+++ meuporg-server/source/Components/StaticMarkerComponent.cpp
@@ -1,6 +1,6 @@
 #include "StaticMarkerComponent.hpp"
 
-StaticMarkerComponent::StaticMarkerComponent(unsigned long long ownerId)
+StaticMarkerComponent::StaticMarkerComponent(unsigned int ownerId)
      : kantan::Component(std::string("StaticMarker"), ownerId)
      , isStatic(false)
 {}
