+++ meuporg-server/source/Components/TowerAIComponent.cpp
@@ -1,6 +1,6 @@
 #include "TowerAIComponent.hpp"
 
-TowerAIComponent::TowerAIComponent(unsigned long long ownerId)
+TowerAIComponent::TowerAIComponent(unsigned int ownerId)
      : kantan::Component(std::string("TowerAI"), ownerId)
 	 , state(TowerAIComponent::TowerAIState::LOOKING_FOR_TARGET)
 	 , target(0)
