+++ meuporg-server/source/Kantan-2/Component/DeletionMarkerComponent.hpp
@@ -12,7 +12,7 @@ namespace kantan
     {
         public:
             // Ctor.
-            DeletionMarkerComponent(unsigned long long ownerId);
+            DeletionMarkerComponent(unsigned int ownerId);
 
             // Attribute.
             bool marked;
