+++ meuporg-server/source/Kantan-2/Component/ParticleEffectComponent.hpp
@@ -28,14 +28,14 @@ namespace kantan
     {
         public:
             // Ctor.
-            ParticleEffectComponent(unsigned long long ownerId);
+            ParticleEffectComponent(unsigned int ownerId);
 
             // Attributes.
             std::function<sf::Color(const Particle&, const sf::Vector2f&)> colorFunction;
             std::function<sf::Vector2f(const Particle&, const sf::Vector2f&)> velocityFunction;
 
             sf::Time maxLifetime;
-            unsigned long long numberOfParticles;
+            unsigned int numberOfParticles;
             std::vector<Particle> particles;
             sf::Vector2f emissionCenter;
     };
