+++ meuporg-server/source/Kantan-2/Entity/Entity.cpp
@@ -4,9 +4,9 @@
 namespace kantan
 {
     /// Static.
-    unsigned long long Entity::m_lastid = 0;
+    unsigned int Entity::m_lastid = 0;
 
-    Entity* Entity::getEntityWithId(unsigned long long id, std::vector<Entity*>& entities)
+    Entity* Entity::getEntityWithId(unsigned int id, std::vector<Entity*>& entities)
     {
         auto itr = std::find_if(entities.begin(), entities.end(), [id](kantan::Entity* e){
                                     return e->getId() == id;
@@ -27,7 +27,7 @@ namespace kantan
 	}
 
     /// Id.
-	unsigned long long Entity::getId() const
+	unsigned int Entity::getId() const
 	{
 		return m_id;
 	}
