+++ meuporg-server/source/Kantan-2/GUI/GUIGraph.hpp
@@ -12,7 +12,7 @@ namespace kantan
 			virtual ~GUIGraph();
 
 			void setSize(float x, float y);
-			void setNumberOfPoints(unsigned long long numberOfPoints);
+			void setNumberOfPoints(unsigned int numberOfPoints);
 			void addValue(float x, float y);
 			void update(sf::Time dt);
 
@@ -22,7 +22,7 @@ namespace kantan
 		protected:
 			std::vector<sf::Vector2f> m_values;
 
-			unsigned long long m_numberOfPoints;
+			unsigned int m_numberOfPoints;
 			sf::Vector2f m_size;
 			sf::VertexArray m_vertexArray;
 	};
